requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
blinker==1.5
blobfile==2.0.1
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
cryptography==39.0.1
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gpt-index==0.4.7
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.1
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
langchain==0.0.83
loguru==0.6.0
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.1.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
rake-nltk==1.0.6
regex==2022.10.31
requests==2.28.2
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
SQLAlchemy==1.4.46
stack-data==0.6.2
streamlit==1.18.1
streamlit-chat==0.0.2.1
tenacity==8.1.0
tiktoken==0.2.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.19.0
wcwidth==0.2.6
weaviate-client==3.11.0
yarl==1.8.2
zipp==3.13.0
